feat: array extensions
changes:
- a lot ;)
- `RandomChoice` method for array extensions added
- unit test added
- a lot of refactoring made

closes #32 ; #33
