+++ src/BB84.Extensions/ArrayExtensions.cs
@@ -1,0 +1,7 @@
+﻿namespace BB84.Extensions;
+
+/// <summary>
+/// The array extensions class.
+/// </summary>
+public static partial class ArrayExtensions
+{ }
