+++ src/BB84.Extensions/DateTimeExtensions.EndOfWeek.cs
@@ -1,0 +1,14 @@
+﻿namespace BB84.Extensions;
+
+public static partial class DateTimeExtensions
+{
+	/// <summary>
+	/// Calculates the last day of the week using the specified date and
+	/// the definition of the first day of the week.
+	/// </summary>
+	/// <param name="dateTime">The date time value to work with.</param>
+	/// <param name="startOfWeek">The first day of the week.</param>
+	/// <returns>The date of the last day of the week.</returns>
+	public static DateTime EndOfWeek(this DateTime dateTime, DayOfWeek startOfWeek = DayOfWeek.Monday)
+		=> StartOfWeek(dateTime, startOfWeek).AddDays(6);
+}
