+++ tests/BB84.ExtensionsTests/ArrayExtensionsTests.RandomChoice.cs
@@ -1,0 +1,14 @@
+﻿namespace BB84.Extensions.Tests;
+
+public sealed partial class ArrayExtensionsTests
+{
+	[TestMethod]
+	public void RandomChoice()
+	{
+		int[] ints = { 1, 2, 3 };
+
+		int i = ints.RandomChoice();
+
+		Assert.IsTrue(ints.Contains(i));
+	}
+}
