+++ tests/BB84.ExtensionsTests/ArrayExtensionsTests.cs
@@ -1,0 +1,5 @@
+﻿namespace BB84.Extensions.Tests;
+
+[TestClass]
+public sealed partial class ArrayExtensionsTests
+{ }
